Add a map system with collision.
Now entities shouldn't be able to walk into each other. Unfortrunately,
this has a bug where if multiple entities try to walk into the same tile
on the same tick, then we don't update the position check in time.

I think the solution is to say that if an entity takes an action, we'll
need to run another pass of action-performing, and only tick initiative
when nobody's done anything.